requirements.txt
binwalk==2.1.0
Flask==2.3.2
librosa==0.10.1
matplotlib==3.7.1
numpy==1.23.5
Pillow==10.2.0
PyExifTool==0.5.6
python_magic==0.4.27
scipy==1.12.0
